# api/gen/http/injective_price_oracle_api/server/server.go
// Code generated by goa v3.7.0, DO NOT EDIT.
//
// Injective Price Oracle API HTTP server
//
// Command:
// $ goa gen github.com/InjectiveLabs/injective-price-oracle/api/design -o ../

package server

import (
	"context"
	"mime/multipart"
	"net/http"

	injectivepriceoracleapi "github.com/InjectiveLabs/injective-price-oracle/api/gen/injective_price_oracle_api"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
	"goa.design/plugins/v3/cors"
)

// Server lists the Injective Price Oracle API service endpoint HTTP handlers.
type Server struct {
	Mounts []*MountPoint
	Probe  http.Handler
	CORS   http.Handler
}

// ErrorNamer is an interface implemented by generated error structs that
// exposes the name of the error as defined in the design.
type ErrorNamer interface {
	ErrorName() string
}

// MountPoint holds information about the mounted endpoints.
type MountPoint struct {
	// Method is the name of the service method served by the mounted HTTP handler.
	Method string
	// Verb is the HTTP method used to match requests to the mounted handler.
	Verb string
	// Pattern is the HTTP request path pattern used to match requests to the
	// mounted handler.
	Pattern string
}

// InjectivePriceOracleAPIProbeDecoderFunc is the type to decode multipart
// request for the "Injective Price Oracle API" service "probe" endpoint.
type InjectivePriceOracleAPIProbeDecoderFunc func(*multipart.Reader, **injectivepriceoracleapi.ProbePayload) error

// New instantiates HTTP handlers for all the Injective Price Oracle API
// service endpoints using the provided encoder and decoder. The handlers are
// mounted on the given mux using the HTTP verb and path defined in the design.
// errhandler is called whenever a response fails to be encoded. formatter is
// used to format errors returned by the service methods prior to encoding.
// Both errhandler and formatter are optional and can be nil.
func New(
	e *injectivepriceoracleapi.Endpoints,
	mux goahttp.Muxer,
	decoder func(*http.Request) goahttp.Decoder,
	encoder func(context.Context, http.ResponseWriter) goahttp.Encoder,
	errhandler func(context.Context, http.ResponseWriter, error),
	formatter func(err error) goahttp.Statuser,
	injectivePriceOracleAPIProbeDecoderFn InjectivePriceOracleAPIProbeDecoderFunc,
) *Server {
	return &Server{
		Mounts: []*MountPoint{
			{"Probe", "POST", "/api/price-oracle/v1/probe"},
			{"CORS", "OPTIONS", "/api/price-oracle/v1/probe"},
		},
		Probe: NewProbeHandler(e.Probe, mux, NewInjectivePriceOracleAPIProbeDecoder(mux, injectivePriceOracleAPIProbeDecoderFn), encoder, errhandler, formatter),
		CORS:  NewCORSHandler(),
	}
}

// Service returns the name of the service served.
func (s *Server) Service() string { return "Injective Price Oracle API" }

// Use wraps the server handlers with the given middleware.
func (s *Server) Use(m func(http.Handler) http.Handler) {
	s.Probe = m(s.Probe)
	s.CORS = m(s.CORS)
}

// Mount configures the mux to serve the Injective Price Oracle API endpoints.
func Mount(mux goahttp.Muxer, h *Server) {
	MountProbeHandler(mux, h.Probe)
	MountCORSHandler(mux, h.CORS)
}

// Mount configures the mux to serve the Injective Price Oracle API endpoints.
func (s *Server) Mount(mux goahttp.Muxer) {
	Mount(mux, s)
}

// MountProbeHandler configures the mux to serve the "Injective Price Oracle
// API" service "probe" endpoint.
func MountProbeHandler(mux goahttp.Muxer, h http.Handler) {
	f, ok := HandleInjectivePriceOracleAPIOrigin(h).(http.HandlerFunc)
	if !ok {
		f = func(w http.ResponseWriter, r *http.Request) {
			h.ServeHTTP(w, r)
		}
	}
	mux.Handle("POST", "/api/price-oracle/v1/probe", f)
}

// NewProbeHandler creates a HTTP handler which loads the HTTP request and
// calls the "Injective Price Oracle API" service "probe" endpoint.
func NewProbeHandler(
	endpoint goa.Endpoint,
	mux goahttp.Muxer,
	decoder func(*http.Request) goahttp.Decoder,
	encoder func(context.Context, http.ResponseWriter) goahttp.Encoder,
	errhandler func(context.Context, http.ResponseWriter, error),
	formatter func(err error) goahttp.Statuser,
) http.Handler {
	var (
		decodeRequest  = DecodeProbeRequest(mux, decoder)
		encodeResponse = EncodeProbeResponse(encoder)
		encodeError    = EncodeProbeError(encoder, formatter)
	)
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ctx := context.WithValue(r.Context(), goahttp.AcceptTypeKey, r.Header.Get("Accept"))
		ctx = context.WithValue(ctx, goa.MethodKey, "probe")
		ctx = context.WithValue(ctx, goa.ServiceKey, "Injective Price Oracle API")
		payload, err := decodeRequest(r)
		if err != nil {
			if err := encodeError(ctx, w, err); err != nil {
				errhandler(ctx, w, err)
			}
			return
		}
		res, err := endpoint(ctx, payload)
		if err != nil {
			if err := encodeError(ctx, w, err); err != nil {
				errhandler(ctx, w, err)
			}
			return
		}
		if err := encodeResponse(ctx, w, res); err != nil {
			errhandler(ctx, w, err)
		}
	})
}

// MountCORSHandler configures the mux to serve the CORS endpoints for the
// service Injective Price Oracle API.
func MountCORSHandler(mux goahttp.Muxer, h http.Handler) {
	h = HandleInjectivePriceOracleAPIOrigin(h)
	mux.Handle("OPTIONS", "/api/price-oracle/v1/probe", h.ServeHTTP)
}

// NewCORSHandler creates a HTTP handler which returns a simple 200 response.
func NewCORSHandler() http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		w.WriteHeader(200)
	})
}

// HandleInjectivePriceOracleAPIOrigin applies the CORS response headers
// corresponding to the origin for the service Injective Price Oracle API.
func HandleInjectivePriceOracleAPIOrigin(h http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		origin := r.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			h.ServeHTTP(w, r)
			return
		}
		if cors.MatchOrigin(origin, "*") {
			w.Header().Set("Access-Control-Allow-Origin", origin)
			w.Header().Set("Vary", "Origin")
			if acrm := r.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				w.Header().Set("Access-Control-Allow-Methods", "POST")
				w.Header().Set("Access-Control-Allow-Headers", "Content-Type")
			}
			h.ServeHTTP(w, r)
			return
		}
		h.ServeHTTP(w, r)
		return
	})
}