# api/gen/http/openapi.yaml
swagger: "2.0"
info:
  title: PriceOracle Service
  description: HTTP server for the Price Oracle API
  version: ""
host: 0.0.0.0:9924
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /api/health/v1/status:
    get:
      tags:
      - health
      summary: GetStatus health
      description: Get current backend health status
      operationId: health#GetStatus
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/HealthGetStatusResponseBody'
            required:
            - s
            - status
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/HealthGetStatusInternalResponseBody'
      schemes:
      - http
  /api/price-oracle/v1/probe:
    post:
      tags:
      - Injective Price Oracle API
      summary: probe Injective Price Oracle API
      description: Validate TOML file
      operationId: Injective Price Oracle API#probe
      consumes:
      - multipart/form-data
      parameters:
      - name: X-Api-Key
        in: header
        description: API key for authentication
        required: false
        type: string
      - name: ProbeRequestBody
        in: formData
        required: true
        schema:
          $ref: '#/definitions/InjectivePriceOracleAPIProbeRequestBody'
          required:
          - content
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/InjectivePriceOracleAPIProbeResponseBody'
            required:
            - result
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/InjectivePriceOracleAPIProbeInvalidArgResponseBody'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/InjectivePriceOracleAPIProbeUnauthorizedResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/InjectivePriceOracleAPIProbeInternalResponseBody'
      schemes:
      - http
      security:
      - api_key_header_X-Api-Key: []
definitions:
  HealthGetStatusInternalResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: Internal Server Error (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  HealthGetStatusResponseBody:
    title: HealthGetStatusResponseBody
    type: object
    properties:
      data:
        $ref: '#/definitions/HealthStatusResponseBody'
      errmsg:
        type: string
        description: Error message.
        example: Something has failed
      s:
        type: string
        description: Status of the response.
        example: error
        enum:
        - ok
        - error
        - no_data
      status:
        type: string
        example: Ut temporibus rem maxime ex eos.
    example:
      data: {}
      errmsg: Something has failed
      s: error
      status: Ut consequatur temporibus non dolor rerum commodi.
    required:
    - s
    - status
  HealthStatusResponseBody:
    title: HealthStatusResponseBody
    type: object
    description: Status defines the structure for health information
    example: {}
  InjectivePriceOracleAPIProbeInternalResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: Internal Server Error (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  InjectivePriceOracleAPIProbeInvalidArgResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: Invalid argument (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  InjectivePriceOracleAPIProbeRequestBody:
    title: InjectivePriceOracleAPIProbeRequestBody
    type: object
    properties:
      content:
        type: string
        description: TOML file contents
        example:
        - 77
        - 111
        - 108
        - 101
        - 115
        - 116
        - 105
        - 97
        - 115
        - 32
        - 117
        - 116
        - 32
        - 112
        - 114
        - 97
        - 101
        - 115
        - 101
        - 110
        - 116
        - 105
        - 117
        - 109
        - 32
        - 112
        - 111
        - 114
        - 114
        - 111
        - 32
        - 110
        - 111
        - 110
        - 32
        - 115
        - 105
        - 110
        - 116
        - 32
        - 110
        - 101
        - 115
        - 99
        - 105
        - 117
        - 110
        - 116
        - 46
        format: byte
    example:
      content:
      - 68
      - 111
      - 108
      - 111
      - 114
      - 101
      - 109
      - 113
      - 117
      - 101
      - 32
      - 113
      - 117
      - 105
      - 97
      - 32
      - 115
      - 101
      - 100
      - 32
      - 101
      - 111
      - 115
      - 46
    required:
    - content
  InjectivePriceOracleAPIProbeResponseBody:
    title: InjectivePriceOracleAPIProbeResponseBody
    type: object
    properties:
      result:
        type: string
        description: Result of the probe
        example: Ea corrupti.
    example:
      result: Inventore non explicabo cumque possimus molestiae enim.
    required:
    - result
  InjectivePriceOracleAPIProbeUnauthorizedResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: Unauthorized (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
securityDefinitions:
  api_key_header_X-Api-Key:
    type: apiKey
    description: |2-

      **Security Scopes**:
        * `api:read`: Read-only access
    name: X-Api-Key
    in: header